[Windows][Performance] Configure MTU in Windows container Container MTU should be properly configured for better performance. MTU cannot be configured with HNSEndpoint on Windows. https://github.com/Microsoft/hcsshim/blob/4a468a6f7ae547974bc32911395c51fb1862b7df/internal/hns/hnsendpoint.go#L12
